Do not send any If header if pseudo locking is enabled.

+++ webdav/src/main/java/ch/cyberduck/core/dav/DAVDeleteFeature.java
@@ -24,6 +24,7 @@ import ch.cyberduck.core.exception.BackgroundException;
 import ch.cyberduck.core.features.Delete;
 import ch.cyberduck.core.features.Lock;
 import ch.cyberduck.core.http.HttpExceptionMappingService;
+import ch.cyberduck.core.preferences.HostPreferences;
 import ch.cyberduck.core.transfer.TransferStatus;
 
 import org.apache.http.HttpHeaders;
@@ -61,10 +62,11 @@ public class DAVDeleteFeature implements Delete {
             deleted.add(file.getKey());
             callback.delete(file.getKey());
             try {
-                if(session.getFeature(Lock.class) != null && file.getValue().getLockId() != null) {
+                if(session.getFeature(Lock.class) != null && file.getValue().getLockId() != null &&
+                        !new HostPreferences(session.getHost()).getBoolean("fs.lock.implementation.pseudo")) {
                     // Indicate that the client has knowledge of that state token
                     session.getClient().delete(new DAVPathEncoder().encode(file.getKey()),
-                        Collections.singletonMap(HttpHeaders.IF, String.format("(<%s>)", file.getValue().getLockId())));
+                            Collections.singletonMap(HttpHeaders.IF, String.format("(<%s>)", file.getValue().getLockId())));
                 }
                 else {
                     session.getClient().delete(new DAVPathEncoder().encode(file.getKey()));
